[frontend] feat(threat-arsenal): fix pr feedback (#4659)
